+++ react/components/ItemQuantity/ItemQuantity.tsx
@@ -9,6 +9,7 @@ const ItemQuantity: StorefrontFunctionComponent<Props> = ({
   warningQuantityThreshold,
   selectedQuantity,
   inputType,
+  showLabel,
 }) => {
   const skuDispatch = useSkuDispatch()
   const productDispatch = useProductDispatch()
@@ -20,6 +21,7 @@ const ItemQuantity: StorefrontFunctionComponent<Props> = ({
       skuDispatch={skuDispatch}
       productDispatch={productDispatch}
       inputType={inputType}
+      showLabel={showLabel}
     />
   )
 }
@@ -29,6 +31,7 @@ interface Props {
   selectedItem: Item
   selectedQuantity: number
   inputType: 'stepper' | 'number'
+  showLabel: boolean
 }
 
 ItemQuantity.schema = {
@@ -51,6 +54,12 @@ ItemQuantity.schema = {
         'ui:widget': 'radio',
       },
       default: 'stepper',
+      isLayout: false,
+    },
+    showLabel: {
+      type: 'boolean',
+      title: 'admin/editor.sku-list.item-quantity.showLabel.title',
+      default: true,
       isLayout: false,
     },
   },