+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Indicated unavailability of SKU highlights
 - Added `showLabel` to `sku-inventory`
 - Added number input support for `sku-quantity-selector`
+- Added `showLabel` to `item-quantity`
 
 ### Changed
 - Changed SKU image carousel to new version